Add flip option for the webcam horizontal direction.
The original code assumes that the orientation of the camera in the
horizontal direction increases with the increasing waveleght. That is
true when you build the device and you can rotate the camera freely.

However if you buy a ready device and the direction is the opposite you
have to have an ability to adjust it in the software.

So this commit adds:

- The f key can now flip the camera image in the horizontal direction
- If image is flipped it's saved as a third line in the callibration

Signed-off-by: Cyril Hrubis <metan@ucw.cz>